Began working on the specs for the `expr` module.
Unfortunately, multimethods cannot be instrumented without conflicts, but I
added the specs anyways for documentation (docstring, etc.) purposes.

I renamed `::valid-operator` spec to just `::operator` and used
`::generic-operator` instead for the looser operator spec, which is also the
default for the corresponding multimethod `op-spec`, so that operators need
not always have dedicated specs.

For the `filter-rems` function, I noticed that it called `find-vars` with
expressions instead of `::variable`s, which is fine for memory FORMs, but not
in general for use-cases of `find-vars`. This is why I created a `find-subexprs`
function, which finds all subexpressions in an expression that match against
a given set of expressions.

I also made some function and option names more consistent. `chained-expr` is
now called `mark-exprs`, which is more accurate (the expression is not really
a chain, since subexpressions are not in a total order) and reflects the
fact that this transforms the expression rather than creating a new kind of
expression. Similarly, `nested-expr` is now called `nest-exprs`.