Completed function specs for expression module. Made some changes regarding `::expression` and `::form` specs, which are more specific now. Although this will decrease performance for predicates that build on those specs, I believe that specs should not be bounded by runtime performance considerations, so predicates can differ and maybe use a different ad-hoc spec with `s/valid?`.
A gotcha with `s/coll-of` bit me as I realized that it is valid on maps too,
which makes sense, since a map is a collection as well. I need to search through
cases where I used that and `s/every` and eventually specify with the option
`kind: sequential?`.

I am unsure about the `cnt->ctx` and `ctx->cnt` functions. Maybe I’ll leave
them as “converter glue” or make them more clear and consistent in terms of api
and semantics.