Completely redesigned the `interpret` function, because it was very unclear and cumbersome to work with in its previous state, which used to substitute the first interpretable expression it encountered from the root up to the leaves but then stopped there. So it had to be called multiple times to be useful.
The new `interpret` just substitutes the root expression and nothing else,
using the specialized `interpret-x` functions or `substitute-expr` for
variables. `interpret-walk` can then be used to substitute every expression
with a depth-first, post-order traversal. But it will only substitute each
expression once. The `…*` variants of both functions will also try to interpret
the results of interpretations until there is nothing left to interpret.
Maybe I can add a count parameter to limit iterations but not sure if that
would be useful.

`interpret` can also get instructions via the environment on which expressions
it should “defocus” (leave uninterpreted). Not sure about the term here, but
I wanted to stay closely aligned with the observer, similar to the `observed`
entries that are produced in simplification. This allows the user to specify
precisely what they don’t want to interpret, which includes specific operators
by including their op-symbol, operators in general with `:ops` or any other
expression like a specific variable.

Added some more tests for `interpret…` functions but they are still very
undertested so I need to take care of that.

I also simplified `simplify-content` and improved `simplify-op` by letting it
call `simplify-content` after interpretation for its default operation, which
is what I wanted anyways.