I have been refactoring the evaluation functions in the `expr` module to simplify the api and make it clearer what return values to expect. `=>` and `=>*` return expression results (constants with `=>` and formDNA with `=>*`). `evaluate` and `eval-all` return maps with result data. For `evaluate`, the result is either a constant or the simplified expression, if it cannot be reduced to a value. For `eval-all`, the result is always a sequence of tuples from an interpretation to a constant value. At first I wanted to turn that into a map, which would actually be more useful as a vdict, but the cost of creating an ordered map seems too high considering that the sequence is already ordered and users can create the map for themselves afterwards.
Furthermore, in the `calc` module, I generalized `dna->quaternary` to take
any sequence of constants and convert it into a number. This was necessary for
`make-compare-dna` (which is now called `make-compare-consts`) to be able to
compare arbitrary sequences of constants, which was motivated by the need to
sort maps with vpoint keys (vdicts).

I also fixed and added some specs and tests. The first parameter of formDNA is
now called `varorder` instead of `vars`, which is much more accurate, although
I am still looking for a better label.

The interpret functions in `expr` can now also take `--focus` set, which works
complementary to `--defocus` and can be used to only interpret specific/selected
expressions, which can also be very useful.

I am also working on a new introductory notebook which is not too overwhelming
with all of my special concepts and things that users would probably not care
about at all, so I hope to focus it on the more essential and practical stuff.