Made significant changes to the API and overall namespace organization.
Each module has its own folder now which can have multiple sub-namespaces to
enable a more flexible organization of code. The interface of a module is a
file in the root directory with the module name. It contains aliases to the
functions I want to expose to users and all docstrings and function specs.
I created a `defapi` util macro to help retain the metadata on aliases.

I am still unsure whether I should put the normal specs definitions in a
separate namespace or put them in the interface files as well (where they are
now). At first, I wanted to decouple specs from the module implementations,
but there are some cases in the `expr.operators` ns where I use them for input
validation of each defined operator in a multimethod. However, it is also
quite annoying having to use long qualified spec keys and I believe they should
belong to the API rather than the implementation, since users might also want
to use them in their code. Maybe the multimethod can be in the API as
well?