Completely refactored the API for the `calc` module and fixed all tests accordingly. Tests will now be splitted if there are separete tests on implementation functions, so they get a corresponding test namespace.
As pointed at in the previous commit, the new API approach is to use wrapper
functions for the implementations. This allows for a much more flexible and
thoughtful API design and decoupling between API and implementation layers.
Although there is a lot of redundance in this approach, it is still DRY because
both layers don’t share the same intention and serve different purposes. They
only accidentally match most cases, because in my original design they were
entangled. I tried to disentangle them by removing some code in the impl. that
only concerns the API or would make it more difficult to maintain and moved
some function definitions completely to the API when they are too shallow
and have more of a convenience function for the user.

For functions that take an options map, I have decided to let them be the
first argument in the argument list. This allows me to fix it in place if
I want to use the function in composition (with `partial`) and gives me more
flexibility with varargs. Most of these functions support lower arities, so
that the user does not have to provide options, but I may make a different
choice if that is not the case and if threading is an important use-case.

This is an ongoing process though and the library is still in alpha stage,
so I may make more changes as I see how all of this plays out in real-world
usage.