Redesigned the API for the `expr` module as well, which was much more tricky due to the internal organization of functions that did not match up very well with a clear organization in the API. It is not perfect by any means, but I think it makes sense now.
Created some new aliases for simplify functions and more explicit names for
some of the functions to make the API clearer.

I also developed textual references for each of the concepts that are much
clearer than the previous descriptions and build on top of each other. They
may change/improve in the future and I may also provide descriptions in the
documentation that are more approachable for beginners who are less familiar
with my terminology.

Specs for implementation functions that are purely for testing are now at the
bottom of the API files. This is because they still depend on Specs defined
in these files and I wanted to keep all specs together. However, I might move
them somewhere else in the future (maybe even in the test files themselves) to
declutter the API.

There are now some failing tests in `expr-test` which I have to debug before
moving on.