Rebuilt the `io` module API to the new structure. This finally concludes my API redesign of formform, which hopefully will be the last one. I decided that `parse-tree` is not needed here and would leak too much implementation details (instaparse) into the API. I also prefer `read-expr` and `print-expr` as simpler names and not even use them as aliases anymore.
Also fixed a few spec-related bugs that caused tests in `io` module to fail.
The one for `::const-int?` was particularly nasty, because there was an
implicit assumption in the code via `==` that it must be an integer, causing
an exception rather than `false` for other types. Should check for more such
cases in my specs.

What I did not recognize earlier is that the default case for the `op-spec`
multimethod was actually problematic since it prevented the `::operator` spec
from failing where an expression is not an operator.

I’ll generally have to clean up and verify some of the specs since I don’t
actually test them very much and haven’t paid much attention to e.g. the
behavior of `s/cat` vs `s/coll-of` while I wrote them.