Added `expr/formDNA-perspectives` as a complement to `calc/dna-perspectives` which does the same thing but while transforming a formDNA expressions into an arrangement of formDNA expressions as its perspectives.
The important thing here was to keep the variable names, since the calc function
would only use digits in permutation and that information gets lost.

Returning an expression could also be useful in equivalence transformations,
since all expressions from formDNA perspectives are equivalent (this is the
basis for `calc/equiv-dna` and `expr/equiv`).

Also added a test for it (more may be needed).