Added dna-get for calc API, tested & spec’d
Added `calc/dna-get` as a convenience function to quickly get a specific value
from a dna (also because it isn’t straightforward that `filter-dna` can do the
same thing with a little more effort).

Improved `calc.core/filter-dna` performance for the fully determined case by
using a quaternary index on the dna vector.