developed new `emul` module for cellular automata + api & specs
While working on the `ff-mindform` component in formform-vis, I realized that I
don’t need to couple the CA logic with any specific visualization method. The CA
can instead be a pure function and iteratively output each next generation, as
an n-dimensional vector of cell-values, where n is the dimension of the
automaton.

Based on that idea, I thought it might be useful to have that functionality
built into formform itself. This allows users to run SelFis, mindFORMs,
lifeFORMs and decisionFORMs on the JVM as well as in the browser, with the
visualization/render method of their choice, or just to process the raw CA data
for analysis, statistics, experiments, etc.

For the API, I aimed for simplicity and flexibility (aligned with how I designed
the other modules of formform). Users can quickly setup a specific preset CA,
such as a mindFORM, but also experiment with new types of CA, by extending the
`sys-ini`, `get-umwelt` or `apply-rule` multimethods with their own
specifications and feeding them (iteratively) into a call to `sys-next`.

I already wrote some useful and common sys-ini methods, but some are still
missing, like the repeated/distributed ball. I also try to transfer them to 2d,
when possible.

The rules methods are relatively simple, since I found that I can decouple them
from different ways to get the ‘umwelt’ (neighborhood). This enables me to mix
and match `get-umwelt` methods with different `apply-rules` methods, allowing
for more flexibility to experiment with new CAs.

I am still unsure where exactly the `decisionFORM` CA differs from the
`lifeFORM` CA other than in the sys-ini (I guess), maybe I’ll have to ask Ralf.