stateful CellularAutomaton class (wip) for better encapsulation
This is meant as an alternative to my previous, more functional approach to step
through a CA using lazy seqs from Clojure’s `iterate`. I wanted something more
object-like that manages its own state and encapsulates methods which access,
return and mutate it. This allows me more control over updates and caching and
potentially yields performance and memory improvements from which a graphical CA
implementation (e.g. in formform.vis) can benefit greatly.

All of this is work-in-progress and subject to change.