redesign of the `emul` api, multimethods → records & protocols
After giving it some thought, I decided to move from multimethods to records and
protocols to model the `ini`, `umwelt` and `rule` data/specifications and the
methods to act on them. This makes it easier to clearly define the different
types, what arguments they need and to compose them freely in other functions.
It provides the data-oriented flexibility and extendability I wanted for my
research while also making the system more robust and (maybe) a little more
performant.

A new addition is what I call “species” records. These define different CA
species such as the SelFi or MindFORM by implementing a `Specifier` protocol
with a `specify-ca` method. In this method, they compose `ini`, `umwelt` and
`rule` records, either hard-wired by design or from their user-provided data to
return another record which specifies an actual CA, given a resolution. This
enables me (and others) to precisely define these types of CA without being too
specific yet.

Although this indirection makes it more cumbersome to just set up a CA (which
can be mitigated by convenience functions), I believe that it is beneficial. I
wanted to separate the type of CA a user may want to define given only the
essential data (such as in case of a SelFi a formDNA for its rules and an ini)
from the more generic but overspecified CA spec that can be given to an iterator
or a state machine to compute its evolution. The former can be more easily
manipulated and be used as a blueprint to generate CAs with specific
resolutions (which I deem not essential for the characteristics of the CA
species itself).

The implementation details of this records & protocols approach are hidden from
users, who will just use functions such as `make-ini`, providing a keyword
instead of a classname to instantiate them with their data. However, they can
extend the system with new types using special `defini`, `defumwelt`, `defrule`
and `defspecies` macros. The macros also register each new record in a map,
along with a (optional) docstring that explains its input fields and behavior.
Users can also (independently) register a spec in the `ini/`, `umwelt/`, `rule/`
and `species/` namespaces for automatic validation in the `make` function that
instantiates the record.

I also changed the `CellularAutomaton` class to only mutate its evolution
instead of being coupled to an iterator. This provides more flexibility with
caching and might be more performant and memory-efficient, but I have to
experiment more so this is all work in progress.