performance optimizations in emule module, macro docstrings
Two more protocols `UmweltOptimized` and `RuleOptimized` can now be implemented
by umwelt and rule records to provide faster implementations of their methods.
The caveat is that they then operate on native array structures instead of
vectors for the generations and umwelt methods output a format that can be
processed faster by their compatible rule methods.

This requires more caution when putting the pieces together but already brought
a huge performance boost to my CA implementation in formform.vis that works with
canvas in the browser. Outside of apps, it is not something users should reach
for in exploratory work, because of the loss in flexibility and compatibility
with formform and other Clojure libraries. It is also only meant for the
state-machine implementation, not the iterator.

This state machine has also been improved by implementing caching of the CA
history up to a certain number of generations, which depends on the resolution
of the CA (to prevent it from eating up too much memory). I’m not sure yet where
I should set the hard limit on how many cells in total can be cached (this could
probably be higher than 8 million). Since it doesn’t store the whole history, it
also doesn’t make sense to implement a “get-evolution” method (maybe
“get-history” would still be useful in some way). Instead, there is now a
`get-current-generation` method.

I have noticed that the `mindform` implementation is much faster in my
formform-vis CA than the `lifeform` and `decisionform` implementations. All of
them have “fast” methods implemented and there is not much difference in terms
of what constructs they use, so I am puzzled why they perform so differently. It
may have something to do with how canvas display operations are optimized in the
browser, but I’ll have to investigate further.