emul module: tests, more general `fill-all`, fixed life CAs, etc.
I added lots of comprehensive test cases for all CA types in formform.emul. They
are compared for congruency to evolution data straight from Ralfs original
demonstrations and thus provide high confidence in the correctness of my
implementations (although arguably many redundant checks here).

The ini `:fill-all` has been extended to take a `pattern` instead of just a
single value. The pattern can be a constant (or `:rand`), an explicit generation
(great for testing or continuing evolutions) or a function of the generation
coordinate/index.

The umwelt `:moore` has been extended such that the order of values in the
umwelt vector can now be optionally set as `:column-first` instead of
`:row-first`. This is apparently the order that Ralf used for his lifeFORM and
decisionFORM rules, given that it now passes all my tests.

The `specify-ca` method now also takes an options map, which is useful to allow
for overwrites of some specifications (e.g. the ini-spec of lifeforms) and I may
also use it for other options in the future.

The `CASystem` protocol has been extended with the `get-cached-history` method,
which returns the cached history/evolution. This method and
`get-current-generation` can now optionally either return the internal array (if
optimization is enabled) or the more compatible evolution/generation vector
data. This is a compromise, since although it leaks internals, the optimized
array can be useful in further processing for better performance (e.g. in my
`ff-automaton` component in formform-vis).