emul: redesigned ini system based on transducers and figure/ground
I completely redesigned the ini system to enable a simpler and more flexible
specification of inis. It is based on the concept of figure vs ground perception
such that “ground inis” always fill up every cell in the first generation while
“figure inis” are positioned in a certain way “on top of” a ground ini.

Multiple figures can be composed to enable things like multiple instances of the
“ball” ini on top of a cyclic ini background. One can even add “holes” (`:_`) to
use the previous/background value instead (which enables e.g. round shapes for
2d inis). This is made possible by having inis provide transducers which can be
stacked together to act as a single vector transformation. Users designing their
own inis can choose to implement this interface or not, but will not need to
care about it unless they want to compose with predefined inis.

Ini transducers/xforms can be implemented in 1d and 2d, but they should map over
a function that takes the ini environment, which provides the width/height of
the ini generation, the x/y coordinates of the current cell, the cell index and
a pseudo-random number generator to create random values. This RNG comes from
`clojure.test.check.random`, is splittable and takes an optional seed, which
enables reproducability of randomized inis. For the predefined inis, the user
can provide the seed in an options map as the first argument to all ini records.

I still try to figure out how to implement an ini that can distribute figure
inis across the generation. I already implemented a repeatable figure ini, but
this one may require preprocessing that involves the width/height of the gen, so
I am not sure if it is easily doable with my new approach.