emul: removed species type, API changes, introduction notebook
`defspecies` and protocol `Specifier` have been removed to simplify the API. For
the most common “species”, there are now just simple constructor functions that
return a `CASpec`. I found that there is no need for a special “CA species”
type, because it doesn’t add any substantial data that users might want to store
separately from what is already defined in `CASpec`. It only creates additional
steps that make CA creation more confusing and annoying. `assoc`ing a label to a
`CASpec` seems sufficient to me to provide a type/identity and it makes things
more flexible when users want to change parts of the spec.

`create-ca` and `ca-iterator` now take a `resolution` argument, because it has
been removed from the `CASpec` record. Although it is an attribute that
determines the behavior of the CA, resolution need not be bundled with the CA
spec but can be documented somewhere else. My decision has the practical reason
that it is common to try different resolutions with the same spec.

Completely redesigned the help system for the “type” constructors. No more
separate `docs` and `params` functions, users can now just type something like
`(make-ini :help)` to see a list of the available inis or `(make-rule :match
:help)` to see docs for a specific rule type. This is so much more convenient.

Added a `distribution` param to `:random` ini. Implementation is still wip.

Added `-opts` params to every `defini`, `defumwelt` and `defrule` type. Makes
things simpler when adding options later on or passing options to internal
records and keeps me from accidentally creating inconsistent specs.

`:von-neumann` and `:moore` umwelt’s now have a `order` param to allow for a
different cell selection order and to be more explicit about it, because this
choice may not be trivial.

I also wrote an introduction to formform.emul with Clerk (still wip), where I
explain how users can work with the module. Custom Clerk viewers and render
functions help visualize the CA output, I put them in separate `formform-render`
and `formform-viewer` namespaces. The `ca-state` notebook is experimental and
will be removed later.

Tests still have to be updated.