changed all uppercase constants/fdna to lowercase, all tests pass
This is the extension of my earlier decision to change the constants/dna
in FORMula notation to lowercase, now applied to formform as a whole. The
reasoning behind this change is to be consistent with how constants are written
in formform, aligned with the spelling in uFORM iFORM, LoF, etc.

The other reason behind it is that my recent change to reverse (or rather
“un-reverse”) formDNA both in FORMula and formform overall now makes it hard to
detect for myself and other users where formDNA hasn’t yet been adapted to the
new notation. This would have bad consequences, even though I believe the
userbase for formform might still just be myself.

I am still unsure if I should keep these modifications.