calc: random fns now take a seed, added weights selection + tests
Random functions are now based on clojure.test.check.random, which I already
used in the emul module. This means they can take a seed for reproducability.

`rand-vpoint` doesn’t work with 0-arity anymore, because the new random
functions don’t return an infinite lazy seq, but I believe that has never been
useful anyway.

Furthermore, `rand-dna` doesn’t take an `elems` vector anymore to select from,
which could have arbitrary elements. I wanted to have the return value be only
formDNA to make it more focussed and maybe add a `rand-dna-seq` or something
later on if other values are really necessary (which I doubt).

I also added variants that take a weights selection, so random values can be
selected with uneven probability. It can be given in the form of a sequence, map
or single number between 0.0 and 1.0 (which is the prob. of `:u`, `:i` and `:m`
against `:n`).