emul: removed :ball ini → ini-patterns, more specs, docs, tests
Removed the `:ball` ini as it is essentially just a collection of patterns for
`:figure`. These patterns are now moved to `ini-patterns` in the emul api.

The `anchor` parameter in various figure-related inis now also takes just a
position/alignment value instead of a map, as a convenience.

Added many more tests for emul api. I don’t think more tests are necessary for
now.

Also additions and fixes for emul api docs, specs and function specs. Plus some
cleanup on comments.