emul: non-essential `:type` and `:kind` data for records
The `i/→rec` wrapper adds `:type` and `:kind` entries to records, so this data
can be easily extracted from the records by the user without having to resort to
implementation details and use converter functions on the tag symbol. Of course,
it is now important that all directly created ini, umwelt and rule records get
created using this wrapper function, but the api functions already use it
internally, so there is no overhead.

The entries get assoc’d on the record after creation, so the record will not
disintegrate when `:type` or `:kind` are removed. It doesn’t make sense to make
them part of the records params, because the same information is already present
in the record tag.