expr: `make-exprs` & `nest-exprs` changed for better function api
I noticed that these functions take option maps while they actually should be
differentiated into multiple functions. For that, I reused my `make` vs `form`
distinction to differentiate between marked/unmarked output and also created
separate functions for left and right nesting. This makes sense, because these
are not just options, but distinct variants of what the user wants to create
with these functions and it is necessary to actively choose between them.

`make-exprs` and `nest-exprs` are still supported, but marked as obsolete.