expr: options last in `eval->expr-all` & `eval-all`, improved docs
`expr/eval->expr-all` (alias `expr/=>*`) and `expr/eval-all` now take `opts` as
their last argument instead of the first one. This is to make argument order of
the options map consistent with other functions in formform and with the
convention in the Clojure community. It works better with the thread-first macro
and is more intuitive to read.

Also improved docstrings of all eval functions.