expr: redesigned eval functions, spec change for ::dna, etc.
This is a massive redesign of all evaluation functions in the `expr` api.

`eval->expr` / `=>` now returns the simplified expression instead of a hole `:_`
for uninterpretable input (such as unregistered symbols or uninterpreted
variables). Likewise, `eval->expr-all` / `=>*` returns the (pre-)simplified,
uninterpreted expression if it could not be determined to a value in any of its
interpretations. It will also just return simple values instead of wrapping them
in a formDNA expr.

The new pre-simplification step can be toggled off by the option `pre-simplify?`
and a new post-reduction step on the resulting formDNA can also be turned off
via `:reduce-dna?`. Another option, `:allow-value-holes?`, re-enables the
previous behavior to return a formDNA expression with holes instead of returning
the simplified expr.

There are now two new functions `eval->val` / `==>` and `eval->val-all` /
`==>*`. The former only returns simple values, including `:_`. The latter
returns only complex values, i.e. raw formDNA (possibly with holes). In essence,
they return output that can be processed by functions in the `calc` as well as
the new `emul` module.

A derived function, `eval-tsds->val-all`, behaves just like `==>*`, but takes a
6-digit binary selection for a TsDS expr., which is especially convenient for
`emul` functions.

In contrast to `=>*`, formDNA by default does not get reduced for “eval->val”
functions and no pre-simplification step applied before interpretation. This
ensures that terms remain consistent and predictable.

Finally, `evaluate` now separates output values in `:result` and the simplified
expr. in `:simplified`. The result is now `nil` if it could not be determined to
a value. `eval-all` now by default returns a map for `:results` instead of a
seq (the option `:ordered-results?` can switch it back if order matters).
Results are const or `nil`, but with the option `:rich-results?` can also be
maps as returned by `evaluate`.

Other changes:

I added warnings to the docstrings of `expr/equal` and `expr/equiv`, because I
would like to reassess the correctness of these functions for comparing
expressions. The calc equivalents work, because they just compare formDNA input,
but expressions are more complex and simplification might have more edge-cases
where two actually equivalent exprs could not be equally well simplified and
thus comparison fails. Maybe one can only be certain when `equiv` returns true
but not otherwise.

Changed the spec for `::dna` to include holes `:_`. This is important to make
incomplete formDNA results work, but I am uncertain about the decision. Maybe a
separate spec would be better?

`expr/expr->const` now also includes identity matches on consts. It makes sense,
because constants are also expressions. Not sure why I published this function
in the api though.