expr: eval-all results changed back to vector of tuples
I figured that my recent change to return a map in `:results` from `eval-all`
was rushed in the process of redesigning the eval apis and I didn’t consider the
inconvenience of having unordered results. Seeing the overall pattern when
scanning results is as important as getting to individual entries.

It is also easier to filter vector results based on the same algorithm that I
use for `filter-dna` in calc (the function actually just works here too).

If the user needs a map instead, they can simply write `(into {} results)`.