ensured correct simplification of holes, fixed symbol simplif.
`expr/simplify` and `expr/simplify-in` now treat each instance of an
expression/value hole (`:_`) in an expression/context as a unique, distinct
symbol. This means that holes in subexpressions do not get simplified by law of
calling like they used to. This is a precaution until I’ve figured out how to
correctly deal with holes in simplification.

Also wrote some tests for hole simplifications.

Fixed a mistake where `expr.core/simplify-content` used to call the method
`interpret-sym` for expression symbols, which was incorrect; it now calls
`simplify-sym`. Since reducers are barely used in expr symbols, I haven’t
noticed it before.

Added explanatory comments to various simplify functions in the implementation,
so outsiders can better understand how the algorithm works and I can refresh my
memory faster when I need to work on it.