expr: value holes throw exc. in simpl., can be processed with option
In simplification and evaluation functions, value holes (`:_`) in the input
expression now cause an exception to be thrown. This makes it clear that they
cannot be processed algebraically.

Via the option `allow-hole-exprs?` they can still be treated like unique
variables, which makes them self-equivalent and therefore valid expressions
again. But this is a “hack”, so I would not like it to be the default,
especially because the distinction/homogenization stages can potentially slow
down evaluation significantly.

For the option to be recognized during simplification, I had to inject it into
the env beforehand. Isolating these options under the `:—opts` key makes sure
that they do not get mixed up with real expressions in the env. Maybe I’ll find
a cleaner solution in the future.

I also distinguished this option from `allow-hole-results?`, because allowing
holes in results is very different from allowing them in expressions during
simplification.

Other changes:

Fixed render and viewer namespaces in notebooks where I called the wrong Clerk
API in render functions, which caused issues with SSR builds (although there are
still other issues here on Clerk’s side).