calc: renamed X-partial-dna-assuming-holes-equal? to X-partial-dna?
Renamed `equal-partial-dna-assuming-holes-equal?` and
`equiv-partial-dna-assuming-holes-equal?` to just `equal-partial-dna?` and
`equiv-partial-dna?`. I decided that the assumption part is unneccessary,
because in order to understand what “partial dna” means, users would have to
read the docstring anyway and use these functions more intentionally, so there
is no risk of not knowing the assumption. Furthermore, it would not even be
possible to compare any partial dna without that assumption.

Also did some cleanup in calc and wrote some more tests related to the
equal/equiv functions and holes situation.