emul: ratio for figure position, changed api for :figure-repeat ini
When specifying a position in figure-related ini specs, users can now also
provide a ratio (a float number in the unit interval as a fraction of the size
of a generation) instead of an index number. This is useful for size-independent
positioning, since relative keywords (`:center`, etc.) may sometimes be too
coarse.

Ini type `:figure-repeat` now takes a figure ini (type `:figure` or
`:rand-figure`) instead of a background ini and a pattern. This increases
flexibility, simplicity and composability for figure repetition. The background
and anchor will of course now be specified in the given figure ini.